@@ -4,7 +4,7 @@ dynamic = ["version"]  # version is managed by build backend
 name = "karlicoss_pymplate"
 dependencies = [
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 
 ## these need to be set if you're planning to upload to pypi
 # description = "TODO"
@@ -33,7 +33,7 @@ testing = [
     "ruff",
     "mypy",
     "lxml", # for mypy html coverage
-    "ty>=0.0.1a15",
+    "ty>=0.0.1a19",
 ]
 
 
